Remove legacy home screen widget view
